KEYCLOAK-13221 Fix arquillian tests on OpenJDK
