+++ tests/arq/src/test/java/org/jboss/test/arquillian/ce/sso/SsoServerMysqlTest.java
@@ -44,7 +44,7 @@ import java.net.URL;
         })
 @OpenShiftResource("https://raw.githubusercontent.com/${template.repository:jboss-openshift}/application-templates/${template.branch:master}/secrets/sso-app-secret.json")
 @OpenShiftResource("https://raw.githubusercontent.com/${template.repository:jboss-openshift}/application-templates/${template.branch:master}/secrets/eap-app-secret.json")
-public class Sso72ServerMysqlTest extends SsoServerTestBase {
+public class SsoServerMysqlTest extends SsoServerTestBase {
     @RouteURL("sso")
     private URL routeURL;
 
@@ -61,6 +61,6 @@ public class Sso72ServerMysqlTest extends SsoServerTestBase {
         return secureRouteURL;
     }
 
-    public Sso72ServerMysqlTest() {
+    public SsoServerMysqlTest() {
     }
 }
