+++ tests/arq/src/test/java/org/jboss/test/arquillian/ce/sso/SsoServerPostgresqlTest.java
@@ -23,26 +23,18 @@
 
 package org.jboss.test.arquillian.ce.sso;
 
-import org.arquillian.cube.openshift.api.OpenShiftHandle;
 import org.arquillian.cube.openshift.api.OpenShiftResource;
 import org.arquillian.cube.openshift.api.Template;
 import org.arquillian.cube.openshift.api.TemplateParameter;
 import org.arquillian.cube.openshift.impl.enricher.RouteURL;
-import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.arquillian.test.api.ArquillianResource;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import java.net.URL;
-import java.util.Collections;
-import java.util.Map;
-
-import static org.junit.Assert.assertTrue;
 
 @RunWith(Arquillian.class)
-@Template(url = "classpath:templates/${template.prefix:sso}-https.json",
-        labels = "application=sso",
+@Template(url = "classpath:templates/${template.prefix:sso}-postgresql.json",
+        labels = "application=sso,component=server",
         parameters = {
                 @TemplateParameter(name = "IMAGE_STREAM_NAMESPACE", value = "${kubernetes.namespace:openshift}"),
                 @TemplateParameter(name = "SSO_ADMIN_USERNAME", value = "admin"),
@@ -51,16 +43,14 @@ import static org.junit.Assert.assertTrue;
                 @TemplateParameter(name = "HTTPS_PASSWORD", value = "mykeystorepass")
         })
 @OpenShiftResource("https://raw.githubusercontent.com/${template.repository:jboss-openshift}/application-templates/${template.branch:master}/secrets/sso-app-secret.json")
-public class Sso72ServerLogTest extends SsoServerTestBase {
+@OpenShiftResource("https://raw.githubusercontent.com/${template.repository:jboss-openshift}/application-templates/${template.branch:master}/secrets/eap-app-secret.json")
+public class SsoServerPostgresqlTest extends SsoServerTestBase {
 
     @RouteURL("sso")
     private URL routeURL;
 
     @RouteURL("secure-sso")
     private URL secureRouteURL;
-
-    @ArquillianResource
-    OpenShiftHandle adapter;
 
     @Override
     protected URL getRouteURL() {
@@ -71,20 +61,4 @@ public class Sso72ServerLogTest extends SsoServerTestBase {
     protected URL getSecureRouteURL() {
         return secureRouteURL;
     }
-
-    @Test
-    @RunAsClient
-    public void testLogs() throws Exception {
-        try {
-            Map<String, String> labels = Collections.singletonMap("application", "sso");
-            String result = adapter.getLog(null, labels);
-
-            assertTrue(result.contains("Deployed \"keycloak-server.war\""));
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw e;
-        }
-    }
-
-
 }
